segmentation-pipeline/pyproject.toml
[project]
name = "tcga-segmentation-inference"
version = "0.1.0"
description = "Histopathology segmentation inference for TCGA whole-slide images"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "numpy>=1.26",
    "pillow>=10.0",
    "rich>=13.0",
    "segmentation-models-pytorch==0.3.1",
    "six",
    "torch>=2.1",
    "typer>=0.15",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "ruff>=0.9",
    "ty",
]

[tool.pytest.ini_options]
pythonpath = ["."]

[tool.ruff]
target-version = "py312"
line-length = 99

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "SIM", # flake8-simplify
    "ANN", # flake8-annotations
    "RUF", # ruff-specific
]
ignore = [
    "ANN401", # dynamically typed expressions (Any) are disallowed
]

[tool.ruff.format]
docstring-code-format = true
